requirements.txt
openai
tk
numpy
pyserial
SpeechRecognition
PyAudio
